perf(ecsm): upgrade to crypto-bigint 0.7.5 + k256 0.14; compile-time Fp params
- crypto-bigint 0.5.5 → 0.7.5, k256 0.13 → 0.14.0-rc.14
- mul_wide → widening_mul, square_wide → widening_square
- to_le_bytes() now returns EncodedUint; add .into() at call sites
- EncodedPoint/FieldElement no longer at k256 root; bridge in curve.rs
  replaced with Sec1Point / FromSec1Point / ToSec1Point
- DynResidue/DynResidueParams → FixedMontyForm/FixedMontyParams in 0.7.5
- Fp in curve.rs and reference_field.rs now use const_monty_params! macro,
  computing Montgomery reduction constants (R², R³, mod_neg_inv) at
  compile time rather than on every field operation

+++ crypto/ecsm/src/tests/reference.rs
@@ -76,7 +76,7 @@ pub fn replay_double_and_add_reference(
     while round >= 0 {
         let (r, next_op) = if op == 0 {
             let r = point_double(&a);
-            let bit = if k.bit_vartime(round as usize) { 1u8 } else { 0u8 };
+            let bit = if k.bit_vartime(round as u32) { 1u8 } else { 0u8 };
             (r, bit)
         } else {
             let r = point_add(&a, g);
